Extract category builder to separate file
