+++ lib/jazzy/source_category.rb
@@ -1,0 +1,78 @@
+require 'jazzy/source_declaration'
+require 'jazzy/config'
+require 'jazzy/jazzy_markdown'
+
+module Jazzy
+  # Category (group, contents) pages generated by jazzy
+  class SourceCategory < SourceDeclaration
+    extend Config.mixin
+
+    def initialize(group, name, abstract, url_name)
+      super()
+      self.type     = SourceDeclaration::Type.overview
+      self.name     = name
+      self.url_name = url_name
+      self.abstract = Markdown.render(abstract)
+      self.children = group
+    end
+
+    # Group root-level docs into custom categories or by type
+    def self.group_docs(docs)
+      custom_categories, docs = group_custom_categories(docs)
+      type_categories, uncategorized = group_type_categories(
+        docs, custom_categories.any? ? 'Other ' : ''
+      )
+      custom_categories + merge_categories(type_categories) + uncategorized
+    end
+
+    def self.group_custom_categories(docs)
+      group = config.custom_categories.map do |category|
+        children = category['children'].flat_map do |name|
+          docs_with_name, docs = docs.partition { |doc| doc.name == name }
+          if docs_with_name.empty?
+            STDERR.puts 'WARNING: No documented top-level declarations match ' \
+                        "name \"#{name}\" specified in categories file"
+          end
+          docs_with_name
+        end
+        # Category config overrides alphabetization
+        children.each.with_index { |child, i| child.nav_order = i }
+        make_group(children, category['name'], '')
+      end
+      [group.compact, docs]
+    end
+
+    def self.group_type_categories(docs, type_category_prefix)
+      group = SourceDeclaration::Type.all.map do |type|
+        children, docs = docs.partition { |doc| doc.type == type }
+        make_group(
+          children,
+          type_category_prefix + type.plural_name,
+          "The following #{type.plural_name.downcase} are available globally.",
+          type_category_prefix + type.plural_url_name,
+        )
+      end
+      [group.compact, docs]
+    end
+
+    # Join categories with the same name (eg. ObjC and Swift classes)
+    def self.merge_categories(categories)
+      merged = []
+      categories.each do |new_category|
+        if existing = merged.find { |c| c.name == new_category.name }
+          existing.children += new_category.children
+        else
+          merged.append(new_category)
+        end
+      end
+      merged
+    end
+
+    def self.make_group(group, name, abstract, url_name = nil)
+      group.reject! { |doc| doc.name.empty? }
+      unless group.empty?
+        SourceCategory.new(group, name, abstract, url_name)
+      end
+    end
+  end
+end
